LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/microtasks/checkpoint-after-window-onerror-module-expected.txt

CONSOLE MESSAGE: Error: script 1

PASS Promise resolved during #report-the-error
PASS Promise resolved during event handlers other than error
